UI improvements
1) common route for UI commons, 2) UI code splits, 3) removed old test client, 4) p2e code splits,5) added marc record sort, 6) test case updates.
