+++ src/clients/common/ui-utils.js
@@ -1,0 +1,48 @@
+//*****************************************************************************
+//
+// Utilities for UI
+//
+//*****************************************************************************
+
+//-----------------------------------------------------------------------------
+
+export function startProcess()
+{
+  const progressbar = document.querySelector(`#progressbar`);
+  progressbar.className="progress-bar";
+}
+
+export function stopProcess()
+{
+  const progressbar = document.querySelector(`#progressbar`);
+  progressbar.className="progress-bar-inactive";
+}
+
+//-----------------------------------------------------------------------------
+
+export function showTab(...tabs) {
+  const root = document.getElementById('root');
+  for (const child of root.children) {
+    if (tabs.includes(child.getAttribute('id'))) {
+      child.hidden = false;
+      child.style.display = null;
+    } else {
+      child.hidden = true;
+      child.style.display = 'none';
+    }
+  }
+}
+
+export function resetForms(...elems) {
+  for (const elem of elems) {
+    const forms = elem.querySelectorAll('form');
+    for (const form of forms) {
+      form.reset();
+    }
+  }
+}
+
+export function reload() {
+  // Programmatically reload page and reset forms
+  location.reload();
+}
