+++ src/clients/muuntaja/muuntaja.html
@@ -10,8 +10,11 @@
   <link rel="stylesheet" href="stylesheets/muuntaja.css" />
   <link rel='stylesheet' href='https://fonts.googleapis.com/css?family=Montserrat' />
   <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
+
   <script src="https://cdn.jsdelivr.net/npm/sortablejs@latest/Sortable.min.js"></script>
-  <script type="text/javascript" src="muuntaja.js"></script>
+
+  <script type="module" src="muuntaja.js"></script>
+  <!-- <script type="text/javascript" src="../common/index.js"></script> -->
 </head>
 
 <!-- -------------------------------------------------------------------------- -->
