+++ src/clients/muuntaja/muuntaja.js
@@ -4,100 +4,18 @@
 //
 //*****************************************************************************
 
-//-----------------------------------------------------------------------------
+import {startProcess, stopProcess} from "../common/ui-utils.js";
+import {showTab, resetForms, reload} from "../common/ui-utils.js";
 
-const RESTurl = window.location.protocol + "//" + window.location.host + "/rest";
-
-console.log(`REST: ${RESTurl}`);
-
-//-----------------------------------------------------------------------------
-// stored auth token
-//-----------------------------------------------------------------------------
-
-const melindaUser = {
-  storage: window.sessionStorage,
-  name: 'melinda-user',
-  get (jsonField = this.name) {
-    try {
-      return JSON.parse(this.storage.getItem(jsonField));
-    } catch (e) {
-      return undefined;
-    }
-  },
-  set (token) {
-    return this.storage.setItem(this.name, JSON.stringify(token));
-  },
-  remove () {
-    return this.storage.removeItem(this.name);
-  }
-};
-
-//-----------------------------------------------------------------------------
-// info needed for muuntaja merge REST call:
-// - Base record
-// - Transform options
-// - Source record
-// - Field selections
-// - User edits
-//-----------------------------------------------------------------------------
-
-var transformed = {
-  profile: null,
-  source: null,
-  base: null,
-  exclude: {},
-  replace: {},
-}
-
-var editmode = false;
-var editing = null;
-
-//-----------------------------------------------------------------------------
-
-function startProcess()
-{
-  const progressbar = document.querySelector(`#progressbar`);
-  progressbar.className="progress-bar";
-}
-
-function stopProcess()
-{
-  const progressbar = document.querySelector(`#progressbar`);
-  progressbar.className="progress-bar-inactive";
-}
-
-function showTab(...tabs) {
-  const root = document.getElementById('root');
-  for (const child of root.children) {
-    if (tabs.includes(child.getAttribute('id'))) {
-      child.hidden = false;
-      child.style.display = null;
-    } else {
-      child.hidden = true;
-      child.style.display = 'none';
-    }
-  }
-}
-
-function resetForms(...elems) {
-  for (const elem of elems) {
-    const forms = elem.querySelectorAll('form');
-    for (const form of forms) {
-      form.reset();
-    }
-  }
-}
-
-function reload() {
-  // Programmatically reload page and reset forms
-  location.reload();
-}
+import {melindaUser, createToken} from "../common/auth.js"
+import {authRequest, profileRequest, transformRequest} from "../common/rest.js";
+import {transformed, showTransformed, stripFieldDecorations} from "../common/marc-record-ui.js";
 
 //-----------------------------------------------------------------------------
 // on page load:
 //-----------------------------------------------------------------------------
 
-function initialize() {
+window.initialize = function() {
   console.log('Initializing');
 
   // Get auth token, if it exists
@@ -126,7 +44,7 @@ function initialize() {
 // Login & logout
 //-----------------------------------------------------------------------------
 
-function login(e) {
+window.login = function(e) {
   e.preventDefault();
 
   console.log('Login:', e);
@@ -143,14 +61,9 @@ function login(e) {
 
   const username = document.getElementById('username').value;
   const password = document.getElementById('password').value;
+  const token = createToken(username, password);
 
-  function generateAuthorizationHeader(username, password = '') {
-    //const encoded = Buffer.from(`${username}:${password}`).toString('base64');
-    const encoded = btoa(`${username}:${password}`);
-    return `Basic ${encoded}`;
-  }
-
-  authRequest(generateAuthorizationHeader(username, password))
+  authRequest(token)
     .then(success)
     .catch(failure);
 
@@ -203,38 +116,7 @@ function authSuccess(user) {
   //stopProcess();
 }
 
-function authRequest(token, url = '') {
-  return fetch(
-    `${RESTurl}/auth/${url}`,
-    {
-      method: 'POST',
-      headers: {
-        Authorization: token
-      }
-    }
-  );
-}
-
-function getAuthToken() {
-  const user = melindaUser.get();
-  if(!user) return;
-  
-  return user.Token;
-}
-
 //-----------------------------------------------------------------------------
-
-function profileRequest() {
-  return fetch(
-    `${RESTurl}/muuntaja/profiles`,
-    {
-      method: 'GET',
-      headers: {
-        Authorization: getAuthToken()
-      }
-    }
-  );
-}
 
 function setProfiles(profiles) {
   console.log("Profiles:", profiles)
@@ -296,13 +178,13 @@ function createMenuBreak() {
 
 //-----------------------------------------------------------------------------
 
-function onNew(e) {
+window.onNew = function(e) {
   console.log('New:', e);
   resetForms(document.getElementById('muuntaja'));
   return eventHandled(e);
 }
 
-function onEdit(e) {
+window.onEdit = function(e) {
   console.log('Edit:', e);
   editmode = !editmode;
   if(editmode) {
@@ -314,340 +196,52 @@ function onEdit(e) {
   return eventHandled(e);
 }
 
-function onSearch(e) {
+window.onSearch = function(e) {
   console.log('Search:', e);
   //const dialog = document.getElementById('searchDlg');
   //console.log('Dialog:', dialog);
   //dialog.show();
 }
 
-function onSave(e) {
+window.onSave = function(e) {
   console.log('Save:', e);
   return eventHandled(e);
 }
 
-function onSettings(e) {
+window.onSettings = function(e) {
   console.log('Settings:', e);
   return eventHandled(e);
 }
 
-function onAccount(e) {
+window.onAccount = function(e) {
   console.log('Account:', e);
   logout();
 }
 
-function ignore(e) {
+window.ignore = function(e) {
   console.log("Ignore")
   return eventHandled(e);
 }
 
-function eventHandled(e) {
+window.eventHandled = function(e) {
   e.stopPropagation();
   e.preventDefault();
   return true;
 }
 
 //-----------------------------------------------------------------------------
-// Exclude/include fields
-//-----------------------------------------------------------------------------
-
-function toggleField(event, field) {
-  const uuid = field.uuid;
-
-  //console.log("Toggle:", uuid)
-
-  if(!transformed.exclude[uuid]) {
-    transformed.exclude[uuid] = stripFieldDecorations(field);
-  } else {
-    delete transformed.exclude[uuid];
-  }
-
-  doTransform();
-}
-
-//-----------------------------------------------------------------------------
-// Field edit
-//-----------------------------------------------------------------------------
-
-function editField(event, field) {
-  // Edit-ohje: https://marc21.kansalliskirjasto.fi/bib/05X-08X.htm#050 
-
-  editing = transformed.transformed.record.fields.find(f => f.uuid == field.uuid);
-
-  console.log("Edit:", editing);
-
-  // Find field from edited fields, if found, fill in data from there
-
-  const dlg = document.querySelector("#fieldEditDlg")
-  //console.log(dlg)
-  dlg.style.display = "flex"
-
-  const content = document.querySelector("#fieldEditDlg #field");
-  content.innerHTML = ""
-  addField(content, editing);
-
-  const tag = document.querySelector("#fieldEditDlg #tag");
-  tag.innerHTML = ""
-  tag.appendChild(createInput('tag', 'tag', field.tag))
-
-  const ind1 = document.querySelector("#fieldEditDlg #ind1");
-  ind1.innerHTML = ""
-  ind1.appendChild(createInput('ind1', 'inds', field.ind1))
-
-  const ind2 = document.querySelector("#fieldEditDlg #ind2");
-  ind2.innerHTML = ""
-  ind2.appendChild(createInput('ind2', 'inds', field.ind2))
-
-  const subfields = document.querySelector("#fieldEditDlg #fieldlist");
-  subfields.innerHTML = ""
-
-  for (const subfield of field.subfields) {
-    subfields.appendChild(createSubfield(subfields, subfield));
-  }
-
-//*
-  Sortable.create(subfields, {
-    ghostClass: 'ghost-row',
-    animation: 50,
-  })
-/**/
-}
-
-function createSubfield(parent, subfield) {
-  const row = document.createElement("div")
-  row.classList.add("subfield")
-  row.appendChild(removeButton());
-  row.appendChild(createInput('code', 'code', subfield.code));
-  row.appendChild(createInput('value', 'value', subfield.value));
-  return row;
-
-  function removeButton() {
-    const btn = document.createElement('button');
-    btn.classList.add("material-icons");
-    btn.innerHTML = "close";
-    btn.addEventListener("click", event => {
-      const state = row.getAttribute("disabled");
-      if(state) {
-        row.removeAttribute("disabled");
-      } else {
-        row.setAttribute("disabled", true);
-      }
-      return true;
-    })
-    return btn;  
-  }
-}
-
-function createInput(name, className, value, editable = true) {
-  const input = document.createElement('span');
-  input.setAttribute('id', name);
-  input.classList.add(className);
-  if(editable) {
-    input.classList.add('editable')
-  }
-  input.innerHTML = value;
-  input.contentEditable = editable;
-  return input;
-}
-
-function onAddField(event) {
-  const subfields = document.querySelector("#fieldEditDlg #fieldlist");
-  subfields.appendChild(createSubfield(subfields, {code: '?', value: '?'}))
-  return eventHandled(event);
-}
-
-function editDlgOK(event) {
-
-  const query = (p) => document.querySelector(p);
-
-  const field = {
-    uuid: editing.uuid,
-    tag: query("#fieldEditDlg #tag #tag").textContent,
-    ind1: query("#fieldEditDlg #ind1 #ind1").textContent,
-    ind2: query("#fieldEditDlg #ind2 #ind2").textContent,
-    subfields: Array.from(query("#fieldEditDlg #fieldlist").childNodes)
-      .filter(e => e.classList.contains("subfield"))
-      .filter(e => !e.getAttribute("disabled"))
-      .map(elem => ({
-        code:  elem.getElementsByClassName("code")[0].textContent,
-        value: elem.getElementsByClassName("value")[0].textContent
-      }))
-  }
-
-  console.log("Edited:", field)
-
-  transformed.replace[field.uuid] = stripFieldDecorations(field);
-  doTransform();
-  
-  return editDlgClose(event);
-}
-
-function editDlgUseOriginal(event) {
-  console.log("Using original.");
-  delete transformed.replace[editing.uuid];
-  doTransform();
-  return editDlgClose(event);
-}
-
-function editDlgClose(event) {
-  const dlg = document.querySelector("#fieldEditDlg")
-  //console.log("Close:", dlg)
-  dlg.style.display = "none"
-  return eventHandled(event);
-}
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-function showTransformed()
-{
-  const {source, base, result} = transformed;
-  showRecord(source, 'source');
-  showRecord(base, 'base');
-  showRecord(result, 'result', editmode = editmode);
-}
-
-function showRecord(data, dest, editmode = false, reference = null) {
-  console.log("Show Record:", data);
-
-  const sourceDiv = document.querySelector(`#muuntaja .record-merge-panel #${dest} #Record`);
-
-  // Clear previous content
-  sourceDiv.innerHTML = '';
-
-  if(!data) return;
-
-  if(data.error) {
-    const error = document.createElement('div');
-    error.classList.add("error")
-    error.innerHTML = data.error;
-    sourceDiv.appendChild(error)
-  }
-  
-  if(data.notes) {
-    const notes = document.createElement('div');
-    notes.classList.add("notes")
-    notes.innerHTML = data.notes;
-    sourceDiv.appendChild(notes)
-  }
-  
-  if(data.record)
-  {
-    const record = data.record;
-
-    if(record.leader) {
-      addField(sourceDiv, {tag: 'LDR', value: record.leader}, editmode);
-    }
-
-    for (const field of record.fields) {
-      const row = addField(sourceDiv, field, editmode);
-      if(field.uuid) {
-        if(editmode) {
-          if(field.subfields) row.addEventListener("click", event => editField(event, field))
-          /* Add here custom field editors */
-        } else {
-          row.addEventListener("click", event => toggleField(event, field))
-        }
-      }
-    }
-  }
-}
-
-//-----------------------------------------------------------------------------
-//-----------------------------------------------------------------------------
-
-function addField(div, field, editmode = false) {
-  //console.log(field)
-  const row = document.createElement('div');
-  row.setAttribute('class', 'row');
-
-  if(transformed.exclude[field.uuid]) {
-    row.classList.add("row-excluded");
-  } else if(transformed.replace[field.uuid]) {
-    row.classList.add("row-replaced");
-  } else if(field.from == "source") {
-    row.classList.add("row-fromSource")
-  } else if(field.from == "base") {
-    row.classList.add("row-fromBase")
-  }
-
-  addTag(row, field.tag);
-  addInd(row, field.ind1, field.ind2);
-
-  if (field.value) {
-    addValue(row, field.value);
-  } else if (field.subfields) {
-    for (const subfield of field.subfields) {
-      addSubfield(row, subfield);
-    }
-  }
-
-  div.appendChild(row);
-  return row;
-
-  //---------------------------------------------------------------------------
-  
-  function addTag(row, value) {
-    row.appendChild(makeSpan('tag', value));
-  }
-
-  function addInd(row, ind1, ind2) {
-    const span = makeSpan('inds');
-    add(span, ind1);
-    add(span, ind2);
-    row.appendChild(span);
-
-    function add(span, ind) {
-      const value = ind && ind.trim() || '&nbsp;';
-      span.appendChild(makeSpan('ind', value));
-    }
-  }
-
-  function addValue(row, value) {
-    row.appendChild(makeSpan('value', value));
-  }
-
-  function addSubfield(row, subfield) {
-    const span = makeSpan('subfield');
-    span.appendChild(makeSubfieldCode(subfield.code));
-    span.appendChild(makeSubfieldData(subfield.value));
-    row.appendChild(span);
-  }
-
-  function makeSubfieldCode(code) {
-    return makeSpan('code', `‡${code}`);
-  }
-
-  function makeSubfieldData(value) {
-    return makeSpan('value', value);
-  }
-}
-
-function makeDiv(className, value) {
-  const div = document.createElement('div');
-  div.setAttribute('class', className);
-  if (value) {
-    div.innerHTML = value;
-  }
-  return div;
-}
-
-function makeSpan(className, value) {
-  const span = document.createElement('span');
-  span.setAttribute('class', className);
-  if (value) {
-    span.innerHTML = value;
-  }
-  return span;
-}
-
-//-----------------------------------------------------------------------------
 // Do transform
 //-----------------------------------------------------------------------------
 
-function doTransform(event = undefined) {
+window.doTransform = function(event = undefined) {
   console.log('Transforming');
   if(event) event.preventDefault();
+
+  //console.log('Source ID:', sourceID);
+  //console.log('Base ID:', baseID);
+  console.log("Transforming:", transformed);
+
+  startProcess();
 
   const sourceID = document.querySelector(`#muuntaja .record-merge-panel #source #ID`).value;
   const baseID = document.querySelector(`#muuntaja .record-merge-panel #base #ID`).value;
@@ -660,29 +254,11 @@ function doTransform(event = undefined) {
     transformed.base = { ID: baseID }
   }
 
-  //console.log('Source ID:', sourceID);
-  //console.log('Base ID:', baseID);
-  console.log("Transforming:", transformed);
-
-  startProcess();
-
-  fetch(
-    `${RESTurl}/muuntaja/transform`,
-    {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json',
-        Authorization: getAuthToken()
-      },
-      body: JSON.stringify(stripDecorations(transformed))
-    }
-  )
+  transformRequest(transformed)
     .then(response => response.json())
     .then(records => {
       console.log('Transformed:', records);
-      transformed = decorateRecords(records)
-      showTransformed();
+      showTransformed(decorateRecords(records));
       stopProcess();
     });
 }
@@ -711,17 +287,6 @@ function stripRecordDecorations(record) {
     }
   } else {
     return record;
-  }
-}
-
-function stripFieldDecorations(f) {
-  return {
-    tag: f.tag,
-    ind1: f.ind1,
-    ind2: f.ind2,
-    value: f.value,
-    subfields: f.subfields,
-    uuid: f.uuid,
   }
 }
