--- src/clients/test/testclient.css
@@ -1,38 +1,0 @@
-body {
-  background-color: lightgray;
-}
-
-#sourceRecord {
-  /*margin: auto; */
-  margin-top: 10px;
-  font-family: 'monospace';
-  white-space: pre-wrap;
-  overflow-x: auto;
-  background-color: #f9f9f9;
-}
-
-#sourceRecord .row {
-  margin-top: 0;
-  display: flex;
-  justify-content: left;
-  align-items: baseline;
-}
-
-.code {
-  display: block;
-  color: #e57373;
-  margin-right: 2px;
-}
-
-.marker {
-  display: block;
-  color: #e57373;
-  margin-left: 2px;
-}
-
-.tag,
-.ind1,
-.ind2 {
-  display: block;
-  color: #448aff;
-}
