--- src/clients/test/testclient.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <script type="text/javascript" src="testclient.js"></script>
-  <link rel="stylesheet" href="testclient.css">
-</head>
-
-<body onload="initialize()">
-  <!-- Submit record ID -->
-  <div style="display: flex; padding-bottom: 8px; border-bottom: 1px solid gray;">      
-    <!-- Record to fetch -->
-    Record ID: <input id="melindaId" style="margin-left: 8pt;"/>
-    <!-- <button onClick="getId()">Fetch</button> -->
-    <div style="flex-grow: 1;"></div>
-    <button onClick="clearRecords()">Clear list</button>
-  </div>
-
-  <!-- List of records fetched -->
-  <div id="sourceRecord"></div>
-</body>
-</html>
