+++ src/muuntaja/config/print-to-e/baseRecord.js
@@ -1,0 +1,80 @@
+
+export const baseRecord = {
+  leader: '00000cam^a22006134i^4500',
+  fields: [
+    {
+      tag: '001',
+      value: '000000000',
+      uuid: '07b7a27e-1fc0-4f30-b181-c54cb4e9d039'
+      //uuid: '' // non-clickable
+    },
+    {
+      tag: '007',
+      value: 'cr^||^||||||||',
+      uuid: '7e290cd1-6bed-4447-8fc3-c1f7e41b760a'
+    },
+
+    //*
+    {
+      tag: '008',
+      value: '^^^^^^s2018^^^^fi^||||^o^^^^^|0|^0|fin|^',
+      uuid: '4bc968f1-9186-4d04-ba09-7537d0c4ee95'
+    },
+
+    /**/
+    /*
+    {
+      tag: '041',
+      ind1: '0',
+      ind2: ' ',
+      subfields: [
+        {
+          code: 'a',
+          value: 'eng'
+        }
+      ],
+      uuid: 'ac44d08d-5532-49bc-b2b9-c863cf930b1a'
+    },
+*/
+    {
+      tag: '337',
+      ind1: ' ',
+      ind2: ' ',
+      subfields: [
+        {
+          code: 'a',
+          value: 'tietokonekäyttöinen'
+        },
+        {
+          code: 'b',
+          value: 'c'
+        },
+        {
+          code: '2',
+          value: 'rdamedia'
+        }
+      ],
+      uuid: '66029c77-30a4-42a2-bf23-6499ca97fe87'
+    },
+    {
+      tag: '338',
+      ind1: ' ',
+      ind2: ' ',
+      subfields: [
+        {
+          code: 'a',
+          value: 'verkkoaineisto'
+        },
+        {
+          code: 'b',
+          value: 'cr'
+        },
+        {
+          code: '2',
+          value: 'rdacarrier'
+        }
+      ],
+      uuid: '55e6067b-8eba-44c1-a53c-293641a40729'
+    }
+  ]
+};
