+++ src/muuntaja/test/f41/01.json
@@ -20,5 +20,26 @@
       }
     ]
   },
-  "base": null
+  "base": {
+    "leader": "00000cam^a22006134i^4500",
+    "fields": [
+      {
+        "tag": "001",
+        "value": "000000000",
+        "uuid": "07b7a27e-1fc0-4f30-b181-c54cb4e9d039"
+      },
+      {
+        "tag": "041",
+        "ind1": "0",
+        "ind2": " ",
+        "subfields": [
+          {
+            "code": "a",
+            "value": "eng"
+          }
+        ],
+        "uuid": "ac44d08d-5532-49bc-b2b9-c863cf930b1a"
+      }
+    ]
+  }
 }
